% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_MVN_compatibility.R
\name{check_MVN_compatibility}
\alias{check_MVN_compatibility}
\title{Check for compatibility with applying multivariate normality}
\usage{
check_MVN_compatibility(dat)
}
\arguments{
\item{dat}{}
}
\description{
Returns a list with \code{dat}, the numeric data, and
\code{dat_nonNum}, the data for which it would be difficult
to apply multivariate normality.
}
