% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVNclean.R
\name{MVNclean}
\alias{MVNclean}
\title{Clean multivariate normal data}
\usage{
MVNclean(
  dat,
  dat_proportion = 0.8,
  reject_alpha = pnorm(-3) * 2,
  groupingVar = "none",
  trim = 0.1,
  ...
)
}
\arguments{
\item{dat}{Data frame of numeric variables. If non-numeric variables (or numeric variables with fewer than 5 unique values) are included, a warning is given which can be ignored; these will be set aside while the cleaning is done and then re-combined before returning.}

\item{dat_proportion}{The proportion of data used for robust covariance estimation. Inversely related to number of shifted outliers.}

\item{reject_alpha}{The chi-squared alpha threshold for identifying outliers (a higher alpha leads to more shifted outliers).}

\item{groupingVar}{Optional vector of length \code{nrow(dat)}. If supplied, the pipeline is applied separately for subsets of the data defined by \code{groupingVar}. This should be used with great caution; while by default medians and MADs are preserved, groups may be transformed separately in ways that lead to unexpected results if group comparisons are subsequently examined in, for example, an ANOVA.}

\item{trim}{Numeric value between 0 and .5. For robustness against outliers, this amount of each of the distribution's tails is excluded when estimating the optimal lambda for the Yeo-Johnson transformation.}

\item{Additional}{arguments; currently unused.}
}
\description{
Applies a cleaning pipeline to a numeric dataset
}
\details{
The pipeline involves: \describe{
\item{\code{\link{YeoJohn}}}{After some proportion of the tail is excluded, the optimal \code{lambda} parameter is found to minimize univariate skew for each variable. This \code{lambda} is then applied to all values of the respective variable.}
\item{\code{\link{RUW}}}{Robust univariate winsorization is applied, shifting all tail values which were beyond \code{reject_alpha} (using \code{\link{median}} and \code{\link{mad}}) toward the \code{median} until they are at the \code{reject_alpha} point in the distribution's tail.}
\item{\code{\link{RMW}}}{Robust multivariate winsorization is applied, shifting multivariate outliers toward the centroid. See also \code{\link{RobMulOutliers}}.}
}

The last step implements the robust covariance estimation and multivariate outlier rejection
approach advocated by Leys et al. (2018; \code{doi.org/10.1016/j.jesp.2017.09.011}).
}
\examples{
iris_clean <- MVNclean(iris)

}
