% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YeoJohn.R
\name{YeoJohn}
\alias{YeoJohn}
\title{Apply optimal Yeo-Johnson transformation to a numeric vector}
\usage{
YeoJohn(x, keepMedian = T, keepMAD = T, trim = 0.05)
}
\arguments{
\item{x}{numeric vector to transform}

\item{keepMedian}{Should the result be shifted so that it has the same median as the input vector?}

\item{keepMAD}{Should the result be scaled so that it has the same dispersion (median absolute deviation) as the input vector?}

\item{trim}{Numeric value between 0 and .5. For robustness against outliers, this amount of each of the distribution's tails is excluded when estimating the optimal lambda for the Yeo-Johnson transformation.}
}
\description{
Finds the Yeo-Johnson transformation that minimizes the skewness of the vector.
Returns the transformed vector with an attribute "lambda" (transformation parameter).
Uses \code{\link[VGAM]{yeo.johnson}}.
}
